+++ nodeconductor/structure/apps.py
@@ -1,7 +1,6 @@
 from __future__ import unicode_literals
 
 from django.apps import AppConfig
-from django.contrib.auth import get_user_model
 from django.db.models import signals
 from django_fsm import signals as fsm_signals
 
@@ -12,15 +11,12 @@ class StructureConfig(AppConfig):
 
     def ready(self):
         from nodeconductor.core.models import CoordinatesMixin
-        from nodeconductor.structure.models import ResourceMixin, \
-            Service, set_permissions_for_model, TagMixin
+        from nodeconductor.structure.models import ResourceMixin, Service, TagMixin
         from nodeconductor.structure import handlers
         from nodeconductor.structure import signals as structure_signals
 
-        User = get_user_model()
         Customer = self.get_model('Customer')
         Project = self.get_model('Project')
-        ProjectGroup = self.get_model('ProjectGroup')
 
         signals.post_save.connect(
             handlers.log_customer_save,
@@ -35,18 +31,6 @@ class StructureConfig(AppConfig):
         )
 
         signals.post_save.connect(
-            handlers.create_customer_roles,
-            sender=Customer,
-            dispatch_uid='nodeconductor.structure.handlers.create_customer_roles',
-        )
-
-        signals.post_save.connect(
-            handlers.create_project_roles,
-            sender=Project,
-            dispatch_uid='nodeconductor.structure.handlers.create_project_roles',
-        )
-
-        signals.post_save.connect(
             handlers.log_project_save,
             sender=Project,
             dispatch_uid='nodeconductor.structure.handlers.log_project_save',
@@ -58,44 +42,8 @@ class StructureConfig(AppConfig):
             dispatch_uid='nodeconductor.structure.handlers.log_project_delete',
         )
 
-        signals.post_save.connect(
-            handlers.create_project_group_roles,
-            sender=ProjectGroup,
-            dispatch_uid='nodeconductor.structure.handlers.create_project_group_roles',
-        )
-
-        signals.pre_delete.connect(
-            handlers.prevent_non_empty_project_group_deletion,
-            sender=ProjectGroup,
-            dispatch_uid='nodeconductor.structure.handlers.prevent_non_empty_project_group_deletion',
-        )
-
-        signals.post_save.connect(
-            handlers.log_project_group_save,
-            sender=ProjectGroup,
-            dispatch_uid='nodeconductor.structure.handlers.log_project_group_save',
-        )
-
-        signals.post_delete.connect(
-            handlers.log_project_group_delete,
-            sender=ProjectGroup,
-            dispatch_uid='nodeconductor.structure.handlers.log_project_group_delete',
-        )
-
-        set_permissions_for_model(
-            User.groups.through,
-            customer_path='group__projectrole__project__customer',
-            project_group_path='group__projectrole__project__project_groups',
-            project_path='group__projectrole__project',
-        )
-
-        set_permissions_for_model(
-            ProjectGroup.projects.through,
-            customer_path='projectgroup__customer',
-        )
-
         # increase nc_user_count quota usage on adding user to customer
-        structure_models_with_roles = (Customer, Project, ProjectGroup)
+        structure_models_with_roles = (Customer, Project)
         for model in structure_models_with_roles:
             name = 'increase_customer_nc_users_quota_on_adding_user_to_%s' % model.__name__
             structure_signals.structure_role_granted.connect(
@@ -106,7 +54,7 @@ class StructureConfig(AppConfig):
 
         # decrease nc_user_count quota usage on removing user from customer
         for model in structure_models_with_roles:
-            name = 'decrease_customer_nc_users_quota_on_adding_user_to_%s' % model.__name__
+            name = 'decrease_customer_nc_users_quota_on_removing_user_from_%s' % model.__name__
             structure_signals.structure_role_revoked.connect(
                 handlers.change_customer_nc_users_quota,
                 sender=model,
@@ -135,18 +83,6 @@ class StructureConfig(AppConfig):
             handlers.log_project_role_revoked,
             sender=Project,
             dispatch_uid='nodeconductor.structure.handlers.log_project_role_revoked',
-        )
-
-        structure_signals.structure_role_granted.connect(
-            handlers.log_project_group_role_granted,
-            sender=ProjectGroup,
-            dispatch_uid='nodeconductor.structure.handlers.log_project_group_role_granted',
-        )
-
-        structure_signals.structure_role_revoked.connect(
-            handlers.log_project_group_role_revoked,
-            sender=ProjectGroup,
-            dispatch_uid='nodeconductor.structure.handlers.log_project_group_role_revoked',
         )
 
         signals.pre_delete.connect(