+++ CHANGELOG.rst
@@ -5,6 +5,13 @@ Release NEXT
 ------------
 - Nothing yet.
 
+Release 0.114.0
+---------------
+- Drop project groups.
+- Implement permission mixin, project and customer permissions.
+- Implement ability to set role deprecation time.
+- Implement endpoint to get history of authorized personnel by period.
+
 Release 0.113.1
 ---------------
 - Fix ActionsMetadata.